exclude nonlinear processees for now

+++ src/generate_mtsp.R
@@ -16,7 +16,7 @@ stocks_files <- file.path(data_files, "world_stock_indexes")
 groups <- c("americas", "asia_and_pacific", "europe", "mea")
 
 # multivariate time-series process characteristics
-functional_forms <- c("linear", "nonlinear")
+functional_forms <- c("linear") # "nonlinear"
 error_term_dists <- c("gaussian", "nongaussian")
 sampling_freq <- c("daily", "monthly")
 Ts <- c(100, 500, 1000, 2000, 3000, 4000, 5000)
@@ -65,49 +65,49 @@ for (g in groups){
     for (ed in error_term_dists){
       for (sf in sampling_freq){
         for (size in Ts){
-          file_name <- paste0(g, "_", size)
+          file_name <- paste(g, size, f, ed, sf, sep = "_")
           
-          # functional form
-          if (f == "linear"){
-            var_sim <- VAR.sim(B = B,
-                               n = size,
+          if (!file.exists(file.path(data_files, "simulations", paste0(file_name, ".npz")))){
+            # functional form
+            if (f == "linear"){
+              var_sim <- VAR.sim(B = B,
+                                 n = size,
+                                 lag = p,
+                                 include = "none")
+            }else if (f == "nonlinear"){
+              tvar_fit <- TVAR(target_ts,
                                lag = p,
+                               dummyToBothRegimes = TRUE,
                                include = "none")
-          }else if (f == "nonlinear"){
-            tvar_fit <- TVAR(target_ts,
-                             lag = p,
-                             dummyToBothRegimes = TRUE,
-                             include = "none")
-            B <- cbind(tvar_fit$coefficients[["Bdown"]],
-                       tvar_fit$coefficients[["Bup"]])
-            var_sim <- TVAR.sim(B = B,
-                                n = size,
-                                lag = p,
-                                mTh = 1,
-                                nthresh = 1,
-                                Thresh = tvar_fit$model.specific$Thresh,
-                                include = "none")
+              B <- cbind(tvar_fit$coefficients[["Bdown"]],
+                         tvar_fit$coefficients[["Bup"]])
+              var_sim <- TVAR.sim(B = B,
+                                  n = size,
+                                  lag = p,
+                                  mTh = 1,
+                                  nthresh = 1,
+                                  Thresh = tvar_fit$model.specific$Thresh,
+                                  include = "none")
+            }
+
+            # error term dist
+            if (ed == "gaussian"){
+              var_sim <- var_sim
+            }else if (ed == "nongaussian"){
+              browser()
+              var_sim <- var_sim + runif(n = dim(var_sim)[1], min = 0, max = 1) # nolint
+            }
+
+            # sampling freq
+            if (sf == "daily"){
+              var_sim <- var_sim
+            }else if (sf == "monthly"){
+              var_sim <- var_sim[seq(from = 1, to = dim(var_sim)[1], by = 20), ] # nolint
+            }
+            
+            np$savez(file.path(data_files, "simulations", file_name),
+                     simulation=var_sim, coefficients=B)
           }
-          file_name <- paste0(file_name, "_", f)
-          
-          # error term dist
-          if (ed == "gaussian"){
-            var_sim <- var_sim
-          }else if (ed == "nongaussian"){
-            var_sim <- var_sim + runif(n = dim(var_sim)[1], min = 0, max = 1) # nolint
-          }
-          file_name <- paste0(file_name, "_", ed)
-          
-          # sampling freq
-          if (sf == "daily"){
-            var_sim <- var_sim
-          }else if (sf == "monthly"){
-            var_sim <- var_sim[seq(from = 1, to = dim(var_sim)[1], by = 20), ] # nolint
-          }
-          file_name <- paste0(file_name, "_", sf)
-          
-          np$savez(file.path(data_files, "simulations", file_name),
-                   simulation=var_sim, coefficients=B)
         }
       }
     }
